Add support for PCRE2
This was needed in a R script that uses the 'tm' library, that fails with the following statement:

In order to enable PCRE2 support to R, the best I found was to use the platform-specific package, so I did not install it through Conda but using official repo.
Also, upgraded OS to Ubuntu Xenial, but a newer Debian than Jessie should do.